requirements.txt
asgiref==3.5.2
Django==3.2.15
djangorestframework==3.14.0
djoser
drf_yasg
pillow
pytz==2022.4
sqlparse==0.4.3
typing-extensions==4.3.0
